server/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
